主模块实现

+++ server/Socket.h
@@ -23,6 +23,7 @@ enum SockAttr {
 	SOCK_ISSERVER = 1,//是否服务器 1表示是 0表示客户端
 	SOCK_ISNONBLOCK = 2,//是否阻塞 1表示非阻塞 0表示阻塞
 	SOCK_ISUDP = 4,//是否为UDP 1表示udp 0表示tcp
+	SOCK_ISIP = 8,//是否为IP协议 1表示IP协议 0表示本地套接字
 };
 
 class CSockParam {
@@ -104,7 +105,8 @@ public:
 	virtual int Close() {
 		m_status = 3;
 		if (m_socket != -1) {
-			if (m_param.attr & SOCK_ISSERVER)
+			if ((m_param.attr & SOCK_ISSERVER) && //服务器
+				((m_param.attr & SOCK_ISIP) == 0))//非IP
 				unlink(m_param.ip);
 			int fd = m_socket;
 			m_socket = -1;
@@ -123,16 +125,16 @@ protected:
 	CSockParam m_param;
 };
 
-class CLocalSocket
+class CSocket
 	:public CSocketBase
 {
 public:
-	CLocalSocket() :CSocketBase() {}
-	CLocalSocket(int sock) :CSocketBase() {
+	CSocket() :CSocketBase() {}
+	CSocket(int sock) :CSocketBase() {
 		m_socket = sock;
 	}
 	//传递析构操作
-	virtual ~CLocalSocket() {
+	virtual ~CSocket() {
 		Close();
 	}
 public:
@@ -141,14 +143,21 @@ public:
 		if (m_status != 0)return -1;
 		m_param = param;
 		int type = (m_param.attr & SOCK_ISUDP) ? SOCK_DGRAM : SOCK_STREAM;
-		if (m_socket == -1)
-			m_socket = socket(PF_LOCAL, type, 0);
+		if (m_socket == -1) {
+			if (param.attr & SOCK_ISIP)
+				m_socket = socket(PF_INET, type, 0);
+			else
+				m_socket = socket(PF_LOCAL, type, 0);
+		}
 		else
 			m_status = 2;//accept来的套接字，已经处于连接状态
 		if (m_socket == -1)return -2;
 		int ret = 0;
 		if (m_param.attr & SOCK_ISSERVER) {
-			ret = bind(m_socket, m_param.addrun(), sizeof(sockaddr_un));
+			if (param.attr & SOCK_ISIP)
+				ret = bind(m_socket, m_param.addrin(), sizeof(sockaddr_in));
+			else
+				ret = bind(m_socket, m_param.addrun(), sizeof(sockaddr_un));
 			if (ret == -1) return -3;
 			ret = listen(m_socket, 32);
 			if (ret == -1)return -4;
@@ -171,10 +180,19 @@ public:
 		if (m_param.attr & SOCK_ISSERVER) {
 			if (pClient == NULL)return -2;
 			CSockParam param;
-			socklen_t len = sizeof(sockaddr_un);
-			int fd = accept(m_socket, param.addrun(), &len);
+			int fd = -1;
+			socklen_t len = 0;
+			if (m_param.attr & SOCK_ISIP) {
+				param.attr |= SOCK_ISIP;
+				len = sizeof(sockaddr_in);
+				fd = accept(m_socket, param.addrin(), &len);
+			}
+			else {
+				len = sizeof(sockaddr_un);
+				fd = accept(m_socket, param.addrun(), &len);
+			}
 			if (fd == -1)return -3;
-			*pClient = new CLocalSocket(fd);
+			*pClient = new CSocket(fd);
 			if (*pClient == NULL)return -4;
 			ret = (*pClient)->Init(param);
 			if (ret != 0) {
@@ -184,7 +202,10 @@ public:
 			}
 		}
 		else {
-			ret = connect(m_socket, m_param.addrun(), sizeof(sockaddr_un));
+			if (m_param.attr & SOCK_ISIP)
+				ret = connect(m_socket, m_param.addrin(), sizeof(sockaddr_in));
+			else
+				ret = connect(m_socket, m_param.addrun(), sizeof(sockaddr_un));
 			if (ret != 0)return -6;
 		}
 		m_status = 2;
@@ -224,3 +245,4 @@ public:
 		return CSocketBase::Close();
 	}
 };
+
